Make Command depend on Environment
As we aim at providing command snippets for environments, we need to
make sure  to intialize Environment first.
